refactored category types with module imports
