+++ src/controllers/stat/playersController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import topPlayersCategory from "../../models/category/topPlayersCategory";
+import { topPlayersCategory } from "../../models/category";
 import { scrapePlayerData } from "../../scraper/stat";
 
 async function topPlayersController(req: Request, res: Response) {
