+++ src/models/category/index.ts
@@ -1,0 +1,8 @@
+const categoryTypes: Array<string> = [
+	"clubs",
+	"players",
+]
+
+export default categoryTypes;
+export { default as topClubsCategory } from "./topClubsCategory";
+export { default as topPlayersCategory } from "./topPlayersCategory";
